% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{gendata}
\alias{gendata}
\title{Generate repeated measures data}
\usage{
gendata(
  beta0 = c(0.5, 0, -0.5),
  n = 300,
  m = rep(3, n),
  fr.form = c("normal", "gamma", "uniform", "t", "mixture.normal",
    "mixture.norm.gamma")[1],
  fxr.form = c("uniform", "uniform-plus-R2", "normal", "normal-mean-R",
    "normal-sd-R")[3],
  latent.variable.type = c("intercept", "slope")[1],
  center.x = FALSE
)
}
\arguments{
\item{beta0}{true fixed effects parameter.}

\item{n}{scalar denoting the number of sujects in the study.}

\item{m}{n-dimensional vector containing the number of repeated measures per subject.}

\item{fr.form}{character indicating the distribution of the latent variable. Options are "normal", "gamma", "uniform",
"t", "mixture.normal" and "mixture.norm.gamma" for a mixture of normal and gamma random variables.}

\item{fxr.form}{character indicating the distribution of the fixed covariates. Options are "uniform",
"uniform-plus-R2" where the covariate is uniformly distributed centered at the square of the latent variable r,
"normal","normal-mean-R" where the covariate has a normal distribution with the mean being the latent variable r,
"normal-sd-R" where the covariate has a normal distribution with the standard deviation being |r|.}

\item{latent.variable.type}{character indicating the type of latent variable generated. Options are "intercept" and "slope".
Default is "intercept".}

\item{center.x}{logical indicator. If TRUE, the fixed covariates are centered at zero.}
}
\description{
Generate repeated measures data
}
