% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgllvm.R, R/main.R
\docType{package}
\name{rgllvm}
\alias{rgllvm}
\title{rgllvm package}
\usage{
rgllvm(y.data, x.data, z.data, n, m, p, beta0 = rep(0, p), family = "binomial")
}
\arguments{
\item{y.data}{(N by 2) matrix of the subject ID (first column) and repeated-measures outcome (second column).
The data is in long format, where each row is one time point per subject and N is the total number of repeated measures for all subjects.}

\item{x.data}{(N by (1+p)) matrix of the subject ID (first column) and p-many repeated-measure covariates that exert a fixed effect in the model.
The data is in long format where each row is one time point per subject and N is the total number of repeated measures for all subjects.}

\item{z.data}{(N by (1+q)) matrix of the subject ID (first column) and q-many repeated-measure covariates that have an associated random effect.
The data is in long format where each row is one time point per subject and N is the total number of repeated measures for all subjects. Currently,
only q=1 is implemented.}

\item{n}{scalar denoting the number of sujects in the study.}

\item{m}{n-dimensional vector containing the number of repeated measures per subject.}

\item{p}{scalar denoting the number of fixed effects in the model.}

\item{beta0}{p-dimensional vector containing the initial estimate of the fixed effect parameters. Default is a p-dimensional vector of zeroes.}

\item{family}{character vector denoting the type of generalized linear model. Code currently only implements family="binomial"
for the logistic model.}
}
\value{
\code{rgllvm} returns a list containing
\itemize{
  \item{beta.est:}{fixed effect estimates from the semiparametric estimator
  that does not make any distributional assumptions on the latent variable.}
  \item{beta.var:}{estimated variances of the fixed effects from the  semiparametric
  estimator that does not make any distributional assumptions on the latent variable}
}
}
\description{
This package provides functions to estimate parameters in a
generalized linear latent variable model where the latent variable
is a random intercept or random slope, and the distribution
of the latent variable is left unknown

Estimates parameters for repeated measures data using a generalized linear latent variable model
where the latent variable is a random slope.
}
\section{Details}{

This function uses a semiparametric approach to estimate parameters for repeated measures data with a
generalized linear latent variable model where the distribution of the latent variable is left unspecified.
}

\examples{
## Random Intercept Example
set.seed(1)
n = 500
m = rep(3,n)
beta0 = c(0.5,0,-0.5)
p = length(beta0)
fr.form = "normal"
fxr.form = "uniform"

## produce long form of the repeated measures data
data.set.out <- gendata(beta0=beta0,n,m=rep(3,n),fr.form,fxr.form,
                       latent.variable.type = "intercept",
                                             center.x=FALSE)

y.data <- data.set.out[,c("family","Y")]
x.data <- data.set.out[,c("family",paste0("X",1:p))]
z.data <- data.set.out[,c("family","Z")]
output <- rgllvm(y.data,x.data,z.data,n,m,p,
beta0=rep(0,p), family="binomial")



## Random Slope Example
set.seed(1)
n = 500
m = rep(3,n)
beta0 = c(0.5,0,-0.5)
p = length(beta0)
fr.form = "normal"
fxr.form = "uniform"

## produce long form of the repeated measures data
data.set.out <- gendata(beta0=beta0,n,m=rep(3,n),fr.form,fxr.form,
                       latent.variable.type = "slope",
                                             center.x=FALSE)

y.data <- data.set.out[,c("family","Y")]
x.data <- data.set.out[,c("family",paste0("X",1:p))]
z.data <- data.set.out[,c("family","Z")]
output <- rgllvm(y.data,x.data,z.data,n,m,p,
beta0=rep(0,p), family="binomial")


}
\references{
Garcia, T.P. and Ma, Y. (2015). Optimal estimator for
logistic model with distribution-free random intercept.
Scandinavian Journal of Statistics, 43, 156-171.

Ma, Y. & Genton, M. G. (2010). Explicit estimating equations for semiparametric generalized
linear latent variable models. Journal of the Royal Statistical Society, Series B 72, 475-495.

Wei, Y., Ma, Y., Garcia, T.P. and Sinha, S. (2019). Consistent estimator
for logistic mixed effect models. The Canadian Journal of Statistics, 47, 140-156. doi:10.1002/cjs.11482.

Garcia, T.P. and Ma, Y. (2015). Optimal estimator for
logistic model with distribution-free random intercept.
Scandinavian Journal of Statistics, 43, 156-171.

Ma, Y. & Genton, M. G. (2010). Explicit estimating equations for semiparametric generalized
linear latent variable models. Journal of the Royal Statistical Society, Series B 72, 475-495.

Wei, Y., Ma, Y., Garcia, T.P. and Sinha, S. (2019). Consistent estimator
for logistic mixed effect models. The Canadian Journal of Statistics, 47, 140-156. doi:10.1002/cjs.11482.
}
