% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgllvm.R, R/main.R
\docType{package}
\name{rgllvm}
\alias{rgllvm}
\title{rgllvm package}
\usage{
rgllvm(y, x, z, n, m, p, beta0 = rep(0, p))
}
\arguments{
\item{y}{xx}

\item{x}{XXX}

\item{z}{xxx}

\item{n}{xxx}

\item{m}{xxxx}

\item{p}{xxx}

\item{beta0}{xxx}
}
\value{
\code{rgllvm} returns a list containing
\itemize{
  \item{beta.est:}{xx}
  \item{beta.var:}{xx}
  \item{beta.mle:}{xx}
  \item{beta.mle.var:}{xx}
  \item{beta.pql:}{xx}
  \item{beta.pql.var:}{xx}
  \item{sigma2.mle:}{xxx}
  \item{sigma2.mle.sd:}{xxx}
}
}
\description{
This package provides functions to estimate parameters in a
generalized linear latent variable model where the latent variable
is a random intercept or random slope, and the distribution
of the latent variable is left unknown

Function estimates parameters in a generalized linear latent variable model
where the latent variable is a random intercept or slope, and the distribution
of the latent variable is left unspecified.
}
\references{
Wei, Y., Ma, Y., Garcia, T.P. and Sinha, S. (2019). Consistent estimator
for logistic mixed effect models. The Canadian Journal of Statistics, 47, 140-156. doi:10.1002/cjs.11482.

Garcia, T.P. and Ma, Y. (2015). Optimal estimator for
logistic model with distribution-free random intercept.
Scandinavian Journal of Statistics, 43, 156-171.
}
